% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeConnection.R
\name{DADAbase.closeConnection}
\alias{DADAbase.closeConnection}
\title{Close the connection to the SQL database}
\usage{
DADAbase.closeConnection()
}
\description{
Purges the incoming and newSeqs tables to reset DADAbase to a state where it can recieve a new dataset. After purging, the connection to the MySQL server is closed.
}
