% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commit.R
\name{DADAbase.commit}
\alias{DADAbase.commit}
\title{Commit new sequences to DADAbase}
\usage{
DADAbase.commit(importMat, knownSeqs = NULL)
}
\description{
Commits the sequences in newSeqs to DADAbase for archiving. Reports number of sequences committed.
}
\examples{
DADAbase.commit(seqtab.import$novelSeqs, seqtab.import$knownSeqs)
}
