% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commitNewSeqs.R
\name{DADAbase.commitNewSeqs}
\alias{DADAbase.commitNewSeqs}
\title{Commit new sequences to DADAbase}
\usage{
DADAbase.commitNewSeqs()
}
\description{
Commits the sequences in newSeqs to DADAbase for archiving. Reports number of sequences committed.
}
