% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAllSeqs.R
\name{DADAbase.getAllSeqs}
\alias{DADAbase.getAllSeqs}
\title{Pull all known sequences from DADAbase}
\usage{
DADAbase.getAllSeqs()
}
\description{
Queries the entire archivedSeqs table and returns all known information.
}
\examples{
all <- DADAbase.getAllSeqs()
}
