% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKnownSeqs.R
\name{DADAbase.getKnownSeqs}
\alias{DADAbase.getKnownSeqs}
\title{Pull matching known sequences from DADAbase}
\usage{
DADAbase.getKnownSeqs()
}
\description{
Queries the archivedSeqs table to pull any matching sequences and stored information that show up in the input dataset.
}
\examples{
knownSeqs <- DADAbase.specifyTaxoKeys("GreenGenes, Silva")
}
