% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNewSeqs.R
\name{DADAbase.getNewSeqs}
\alias{DADAbase.getNewSeqs}
\title{Pull new sequences}
\usage{
DADAbase.getNewSeqs()
}
\description{
Queries DADAbase to report any sequence variants not already stored in archivedSeqs.
}
