% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSeqs.R
\name{DADAbase.importSeqs}
\alias{DADAbase.importSeqs}
\title{Close the connection to the SQL database}
\usage{
DADAbase.importSeqs(incomingTaxa)
}
\arguments{
\item{incomingTaxa}{A data.frame coming from the end of the DADA2 pipeline where rows are sequences and the columns are taxonomic levels describing each sequence variant.}
}
\description{
Accepts a taxonomy table and partitions out sequences that don't exist in the archivedSeqs portion of DADAbase.
}
\examples{
DADAbase.importSeqs(taxa)
}
