% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knownSeqsSource.R
\name{DADAbase.knownSeqsSource}
\alias{DADAbase.knownSeqsSource}
\title{Pull the runs and groups that a sequence was found in}
\usage{
DADAbase.knownSeqsSource(key, queryOnSeq = FALSE)
}
\arguments{
\item{key}{The accession number or sequence the}
}
\description{
Searches DADAbase to pull run and group information for a particular sequence. The sequence can be queried on the actual sequence string itself or it can be queried on the accession number of the sequence.
}
\examples{
DADAbase.knownSeqsSource(12)
}
