% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openConnection.R
\name{DADAbase.openConnection}
\alias{DADAbase.openConnection}
\title{Open a connection to the SQL database}
\usage{
DADAbase.openConnection()
}
\description{
Uses the \code{RSQLite} package to establish a connection to DADAbase from the active R session.
}
\examples{
DADAbase.openConnection()
}
