% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareSeqs.R
\name{DADAbase.prepareSeqs}
\alias{DADAbase.prepareSeqs}
\title{Creates a data frame with sequences for annotation}
\usage{
DADAbase.prepareSeqs(seqMatrix, knownMatrix = NULL)
}
\arguments{
\item{seqMatrix}{A data.frame coming from the output of DADAbase.removeKnownSeqs() -- holds novel variants}

\item{knownMatrix}{A data.frame coming from the output of DADAbase.getKnownSeqs() -- holds known variants}
}
\description{
Accepts a sequence matrix and prepares a sequence data frame with predefined columns for import into DADAbase downstream. If knownMatrix is not NULL, then the output will also contain a data frame of known sequences to update DADAbase to show that variant was also found in this run.
}
\examples{
seqtab.import <- DADAbase.prepareSeqs(seqtab.removed, seqtab.known)
}
