% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{DADAbase.query}
\alias{DADAbase.query}
\title{Query DADAbase with string matching}
\usage{
DADAbase.query(queryTerm, matchOn, exactMatching = FALSE)
}
\arguments{
\item{queryTerm}{The term to query for}

\item{matchOn}{The attribute to query on}

\item{exactMatching}{Whether or not the returned sequences should be those that exactly match the queryTerm}
}
\description{
Querys DADAbase by using string matching
}
\examples{
DADAbase.query('V4', 'primers', TRUE)
}
