% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeKnownSeqs.R
\name{DADAbase.removeKnownSeqs}
\alias{DADAbase.removeKnownSeqs}
\title{Pull matching known sequences from a sequence matrix}
\usage{
DADAbase.removeKnownSeqs(seqMatrix, knownData)
}
\value{
A data.frame of sequences and associated taxonomic data from the archive
}
\description{
Queries the archivedSeqs table to pull any matching sequences that show up in the archive already. Uses these matches to remove columns from the sequence matrix so that sequences already annotated are excluded from DADA2's downstream annotation.
}
\examples{
seqtab.removed <- DADAbase.removeKnownSeqs(seqtab.nochim, seqtab.known)
}
