% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifyGroupNums.R
\name{DADAbase.specifyGroupNums}
\alias{DADAbase.specifyGroupNums}
\title{Update the group number column for new sequences}
\usage{
DADAbase.specifyGroupNums(groupNumTable)
}
\arguments{
\item{groupNumTable}{A data frame of sequences and corresponding grouping numbers to be added to new sequences. Specifically should be of class data.frame with two columns titled "Sequence" and "Group". groupNumTable$Group must be of class numeric().}
}
\description{
After import, new sequences have taxonomy and taxonomic keywords fields default to NULL. This function is used to specify values to fill in the taxonomy fields.
}
\examples{
DADAbase.specifyTaxoKeys(" GreenGenes, Silva")
}
