% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifyMetadata.R
\name{DADAbase.specifyMetadata}
\alias{DADAbase.specifyMetadata}
\title{Add metadata to an entry to DADAbase}
\usage{
DADAbase.specifyMetadata()
}
\description{
Stores DADAbase about an import of data into DADAbase based on prompt-based inputs.
}
\examples{
DADAbase.specityMetadata()
}
