% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifyTaxoKeys.R
\name{DADAbase.specifyTaxoKeys}
\alias{DADAbase.specifyTaxoKeys}
\title{Update the taxonomic keywords column for new sequences}
\usage{
DADAbase.specifyTaxoKeys(keys)
}
\arguments{
\item{keys}{Keywords to be added to new sequences to specify annotation tool information.}
}
\description{
After import, new sequences are partitioned out and the taxonomic keywords fields default to "Unspecified". This function is used to specify values to fill in these fields.
}
\examples{
DADAbase.specifyTaxoKeys("GreenGenes, Silva")
}
