% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/specifyTaxonomy.R
\name{DADAbase.specifyTaxonomy}
\alias{DADAbase.specifyTaxonomy}
\title{Update the taxonomy column for new sequences}
\usage{
DADAbase.specifyTaxonomy(taxoTable)
}
\arguments{
\item{taxoTable}{A data frame of sequences and paired taxonomic information to be added to new sequences. Specifically should be of class data.frame with two columns titled "Sequence" and "Taxonomy".}
}
\description{
After import, new sequences have taxonomy and taxonomic keywords fields default to NULL. This function is used to specify values to fill in the taxonomy fields.
}
\examples{
DADAbase.specifyTaxoKeys(" GreenGenes, Silva")
}
