% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importSeqs.R
\name{DADAbase.importSeqs}
\alias{DADAbase.importSeqs}
\title{Close the connection to the SQL database}
\usage{
DADAbase.importSeqs(seqMatrix)
}
\arguments{
\item{seqMatrix}{A data.frame coming from the end of the DADA2 pipeline where rows are samples and the columns are sequence variants.}
}
\description{
Accepts a sequence matrix and partitions out sequences that don't exist in the archivedSeqs portion of DADAbase.
}
\examples{
DADAbase.importSeqs(taxa)
}
